Cleaned up standings API
Standings had trophy and donations included in order to render those on the standings page.  I was not entirely pleased with that, realized how I could solve it after a rubber ducking session with Dreamifi.
Moved trophies to their own API, so first standings are loaded, and then trophies are applied.